requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==5.1.0
Django==4.2.11
django-allauth==0.61.1
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==24.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
